# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Distributed Job Scheduler API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  createdAt: string | null;
}

export interface UserInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Organization {
  id: string;
  name: string;
  /** @nullable */
  ownerUserId?: string | null;
  /** @nullable */
  createdAt: string | null;
}

export interface OrganizationInput {
  /** @minLength 1 */
  name: string;
}

export interface OrgMember {
  orgId: string;
  userId: string;
  role: string;
  /** @nullable */
  joinedAt: string | null;
}

export type OrgMemberInputRole = typeof OrgMemberInputRole[keyof typeof OrgMemberInputRole];


export const OrgMemberInputRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface OrgMemberInput {
  userId: string;
  role: OrgMemberInputRole;
}

export interface Project {
  id: string;
  orgId: string;
  name: string;
  /** @nullable */
  createdAt: string | null;
}

export interface ProjectInput {
  orgId: string;
  /** @minLength 1 */
  name: string;
}

export type RetryPolicyType = typeof RetryPolicyType[keyof typeof RetryPolicyType];


export const RetryPolicyType = {
  fixed: 'fixed',
  linear: 'linear',
  exponential: 'exponential',
} as const;

export interface RetryPolicy {
  id: string;
  type: RetryPolicyType;
  baseDelayMs: number;
  maxRetries: number;
  /** @nullable */
  multiplier?: number | null;
  /** @nullable */
  createdAt?: string | null;
}

export interface Queue {
  id: string;
  projectId: string;
  name: string;
  priority: number;
  concurrencyLimit: number;
  isPaused: boolean;
  /** @nullable */
  defaultRetryPolicyId?: string | null;
  /** @nullable */
  createdAt: string | null;
}

export interface QueueInput {
  projectId: string;
  /** @minLength 1 */
  name: string;
  priority?: number;
  concurrencyLimit?: number;
  defaultRetryPolicyId?: string;
}

export interface QueueUpdate {
  priority?: number;
  concurrencyLimit?: number;
  /** @nullable */
  defaultRetryPolicyId?: string | null;
}

export interface QueueStats {
  queueId: string;
  queued: number;
  running: number;
  completed: number;
  failed: number;
  deadLetter: number;
  total: number;
}

export type JobPayload = { [key: string]: unknown };

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  queued: 'queued',
  scheduled: 'scheduled',
  claimed: 'claimed',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  dead_letter: 'dead_letter',
} as const;

export interface Job {
  id: string;
  queueId: string;
  type: string;
  payload: JobPayload;
  status: JobStatus;
  priority: number;
  runAt: string;
  /** @nullable */
  cronExpression?: string | null;
  /** @nullable */
  retryPolicyId?: string | null;
  attemptCount: number;
  maxAttempts: number;
  /** @nullable */
  idempotencyKey?: string | null;
  /** @nullable */
  workerId?: string | null;
  /** @nullable */
  claimedAt?: string | null;
  /** @nullable */
  createdAt: string | null;
  /** @nullable */
  updatedAt: string | null;
}

export type JobInputPayload = { [key: string]: unknown };

export interface JobInput {
  queueId: string;
  type: string;
  payload?: JobInputPayload;
  priority?: number;
  runAt?: string;
  cronExpression?: string;
  retryPolicyId?: string;
  maxAttempts?: number;
  idempotencyKey?: string;
}

export interface JobPage {
  data: Job[];
  total: number;
  page: number;
  limit: number;
}

export type JobExecutionStatus = typeof JobExecutionStatus[keyof typeof JobExecutionStatus];


export const JobExecutionStatus = {
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type JobExecutionResult = { [key: string]: unknown } | null;

export interface JobExecution {
  id: string;
  /** @nullable */
  jobId?: string | null;
  /** @nullable */
  workerId?: string | null;
  attemptNumber: number;
  status: JobExecutionStatus;
  /** @nullable */
  startedAt: string | null;
  /** @nullable */
  finishedAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  result?: JobExecutionResult;
}

export type JobLogLogLevel = typeof JobLogLogLevel[keyof typeof JobLogLogLevel];


export const JobLogLogLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
} as const;

export interface JobLog {
  id: string;
  /** @nullable */
  jobExecutionId?: string | null;
  /** @nullable */
  timestamp: string | null;
  logLevel: JobLogLogLevel;
  message: string;
}

export type WorkerStatus = typeof WorkerStatus[keyof typeof WorkerStatus];


export const WorkerStatus = {
  active: 'active',
  draining: 'draining',
  dead: 'dead',
} as const;

export interface Worker {
  id: string;
  name: string;
  status: WorkerStatus;
  /** @nullable */
  lastHeartbeatAt?: string | null;
  /** @nullable */
  registeredAt: string | null;
}

export interface WorkerHeartbeat {
  id: string;
  /** @nullable */
  workerId: string | null;
  /** @nullable */
  timestamp: string | null;
  /** @nullable */
  cpuUsage?: number | null;
  /** @nullable */
  memoryUsage?: number | null;
  /** @nullable */
  activeJobCount?: number | null;
}

export type ScheduledJobJobTemplate = { [key: string]: unknown };

export interface ScheduledJob {
  id: string;
  queueId: string;
  jobTemplate: ScheduledJobJobTemplate;
  cronExpression: string;
  /** @nullable */
  nextRunAt?: string | null;
  /** @nullable */
  lastRunAt?: string | null;
  isActive: boolean;
  /** @nullable */
  createdAt: string | null;
}

export type ScheduledJobInputJobTemplate = { [key: string]: unknown };

export interface ScheduledJobInput {
  queueId: string;
  jobTemplate: ScheduledJobInputJobTemplate;
  cronExpression: string;
}

export type DeadLetterJobPayload = { [key: string]: unknown };

export interface DeadLetterJob {
  id: string;
  /** @nullable */
  originalJobId?: string | null;
  queueId: string;
  payload: DeadLetterJobPayload;
  /** @nullable */
  failureReason?: string | null;
  /** @nullable */
  movedAt: string | null;
}

export interface DeadLetterPage {
  data: DeadLetterJob[];
  total: number;
  page: number;
  limit: number;
}

export interface ThroughputPoint {
  time: string;
  completed: number;
  failed: number;
}

export interface ThroughputMetrics {
  range: string;
  data: ThroughputPoint[];
}

export interface MetricsSummary {
  totalJobs: number;
  activeWorkers: number;
  queued: number;
  running: number;
  completed: number;
  failed: number;
  deadLetter: number;
  successRate: number;
}

export type ListProjectsParams = {
orgId: string;
};

export type ListQueuesParams = {
projectId: string;
};

export type ListJobsParams = {
queueId?: string;
status?: string;
page?: number;
limit?: number;
};

export type ListScheduledJobsParams = {
queueId?: string;
};

export type ListDeadLetterJobsParams = {
queueId?: string;
page?: number;
limit?: number;
};

export type GetThroughputMetricsParams = {
queueId?: string;
range?: GetThroughputMetricsRange;
};

export type GetThroughputMetricsRange = typeof GetThroughputMetricsRange[keyof typeof GetThroughputMetricsRange];


export const GetThroughputMetricsRange = {
  '1h': '1h',
  '6h': '6h',
  '24h': '24h',
  '7d': '7d',
} as const;