# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Distributed Job Scheduler API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /orgs:
    get:
      operationId: listOrgs
      tags: [orgs]
      summary: List organizations for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrg
      tags: [orgs]
      summary: Create organization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /orgs/{orgId}/members:
    post:
      operationId: addOrgMember
      tags: [orgs]
      summary: Add a member to organization
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgMember"
        "404":
          description: Org or user not found
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs with filters
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: queueId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPage"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
  /jobs/{jobId}/executions:
    get:
      operationId: getJobExecutions
      tags: [jobs]
      summary: Get execution history for a job
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobExecution"
  /jobs/{jobId}/logs:
    get:
      operationId: getJobLogs
      tags: [jobs]
      summary: Get logs for a job
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobLog"
  /jobs/{jobId}/retry:
    post:
      operationId: retryJob
      tags: [jobs]
      summary: Retry a failed or dead-lettered job
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /dlq:
    get:
      operationId: listDeadLetterJobs
      tags: [dlq]
      summary: List dead letter jobs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: queueId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeadLetterPage"
  /dlq/{dlqId}/requeue:
    post:
      operationId: requeueDeadLetter
      tags: [dlq]
      summary: Requeue a dead letter job
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: dlqId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
    UserInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    Organization:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        ownerUserId:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
    OrganizationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    OrgMember:
      type: object
      required: [orgId, userId, role, joinedAt]
      properties:
        orgId:
          type: string
        userId:
          type: string
        role:
          type: string
        joinedAt:
          type: ["string", "null"]
    OrgMemberInput:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [admin, member]
...
    Job:
      type: object
      required: [id, queueId, type, payload, status, priority, runAt, attemptCount, maxAttempts, createdAt, updatedAt]
      properties:
        id:
          type: string
        queueId:
          type: string
        type:
          type: string
        payload:
          type: object
        status:
          type: string
          enum: [queued, scheduled, claimed, running, completed, failed, dead_letter]
        priority:
          type: integer
        runAt:
          type: string
        cronExpression:
          type: ["string", "null"]
        retryPolicyId:
          type: ["string", "null"]
        attemptCount:
          type: integer
        maxAttempts:
          type: integer
        idempotencyKey:
          type: ["string", "null"]
        workerId:
          type: ["string", "null"]
        claimedAt:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
        updatedAt:
          type: ["string", "null"]
    JobInput:
      type: object
      required: [queueId, type]
      properties:
        queueId:
          type: string
        type:
          type: string
        payload:
          type: object
        priority:
          type: integer
        runAt:
          type: string
        cronExpression:
          type: string
        retryPolicyId:
          type: string
        maxAttempts:
          type: integer
        idempotencyKey:
          type: string
    JobPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    JobExecution:
      type: object
      required: [id, attemptNumber, status, startedAt]
      properties:
        id:
          type: string
        jobId:
          type: ["string", "null"]
        workerId:
          type: ["string", "null"]
        attemptNumber:
          type: integer
        status:
          type: string
          enum: [running, completed, failed]
        startedAt:
          type: ["string", "null"]
        finishedAt:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        result:
          type: ["object", "null"]
    JobLog:
      type: object
      required: [id, logLevel, message, timestamp]
      properties:
        id:
          type: string
        jobExecutionId:
          type: ["string", "null"]
        timestamp:
          type: ["string", "null"]
        logLevel:
          type: string
          enum: [info, warn, error]
        message:
          type: string
...
    DeadLetterJob:
      type: object
      required: [id, queueId, payload, movedAt]
      properties:
        id:
          type: string
        originalJobId:
          type: ["string", "null"]
        queueId:
          type: string
        payload:
          type: object
        failureReason:
          type: ["string", "null"]
        movedAt:
          type: ["string", "null"]
    DeadLetterPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DeadLetterJob"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...