# lib/api-zod/src/generated/types/deadLetterPage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Distributed Job Scheduler API
 * OpenAPI spec version: 0.1.0
 */
import type { DeadLetterJob } from './deadLetterJob';

export interface DeadLetterPage {
  data: DeadLetterJob[];
  total: number;
  page: number;
  limit: number;
}