# lib/api-zod/src/generated/types/orgMemberInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Distributed Job Scheduler API
 * OpenAPI spec version: 0.1.0
 */
import type { OrgMemberInputRole } from './orgMemberInputRole';

export interface OrgMemberInput {
  userId: string;
  role: OrgMemberInputRole;
}