# lib/api-zod/src/generated/types/organizationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Distributed Job Scheduler API
 * OpenAPI spec version: 0.1.0
 */

export interface OrganizationInput {
  /** @minLength 1 */
  name: string;
}