# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Distributed Job Scheduler API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  createdAt: string | null;
}